# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: PriceGuard AI Core Risk API
  version: 0.1.0-mvp
  description: >
    Draft contract for the Phase 2 MVP. This is hand-authored now to fix the contract
    before implementation; `apps/api` also generates a live spec via @nestjs/swagger at
    `/api/docs-json` which must be diffed against this file in CI once both exist
    (tracked follow-up, not yet wired).
servers:
  - url: https://api.priceguard.example/v1
    description: Placeholder — no production domain registered yet.
security:
  - ApiKeyAuth: []
paths:
...
  /appeals/{id}/decision:
    post:
      summary: Record a human decision on an appeal (uphold/overturn)
      operationId: decideAppeal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outcome]
              properties:
                outcome:
                  type: string
                  enum: [UPHELD, OVERTURNED]
                notes:
                  type: string
      responses:
        '200':
          description: OK
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-PriceGuard-Api-Key
...